Add clarifying paragraph to encryption page
I wouldn't want to mislead people into thinking that they can prevent their game's save files from being edited and shared by encrypting and decrypting them locally on the player's own computer. It can't prevent that, the only way to prevent the files from being edited is to keep them on a remote server somewhere, and also run the game logic on that server instead of the local computer.